algorithm/requirements.txt
numpy==1.19.5
pillow==8.1.0
opencv-python==4.5.3.56
matplotlib==3.3.4
cmapy==0.6.6
imgaug==0.4.0
tqdm==4.61.2
prettytable